+++ commands/transfer_funds.go
@@ -76,6 +76,13 @@ func checkTransfer(cmd *commandspb.Transfer) Errors {
 			if k.OneOff.GetDeliverOn() < 0 {
 				errs.AddForProperty("transfer.kind.deliver_on", ErrMustBePositiveOrZero)
 			}
+			// do not allow for one off transfer to one of the metric based accounts
+			if cmd.ToAccountType == vega.AccountType_ACCOUNT_TYPE_REWARD_LP_RECEIVED_FEES ||
+				cmd.ToAccountType == vega.AccountType_ACCOUNT_TYPE_REWARD_MAKER_RECEIVED_FEES ||
+				cmd.ToAccountType == vega.AccountType_ACCOUNT_TYPE_REWARD_TAKER_PAID_FEES ||
+				cmd.ToAccountType == vega.AccountType_ACCOUNT_TYPE_REWARD_MARKET_PROPOSERS {
+				errs.AddForProperty("transfer.account.to", errors.New("transfers to metric-based reward accounts must be recurring transfers that specify a distribution metric"))
+			}
 		case *commandspb.Transfer_Recurring:
 			if k.Recurring.EndEpoch != nil && *k.Recurring.EndEpoch <= 0 {
 				errs.AddForProperty("transfer.kind.end_epoch", ErrMustBePositive)
@@ -102,10 +109,11 @@ func checkTransfer(cmd *commandspb.Transfer) Errors {
 					cmd.ToAccountType == vega.AccountType_ACCOUNT_TYPE_REWARD_MARKET_PROPOSERS) {
 					errs.AddForProperty("transfer.kind.dispatch_strategy", ErrIsNotValid)
 				}
-				// check asset for metric is passed
-				if len(k.Recurring.DispatchStrategy.AssetForMetric) <= 0 {
+				// check asset for metric is passed unless it's a market proposer reward
+				if len(k.Recurring.DispatchStrategy.AssetForMetric) <= 0 && cmd.ToAccountType != vega.AccountType_ACCOUNT_TYPE_REWARD_MARKET_PROPOSERS {
 					errs.AddForProperty("transfer.kind.dispatch_strategy.asset_for_metric", ErrUnknownAsset)
-				} else if !IsVegaPubkey(k.Recurring.DispatchStrategy.AssetForMetric) {
+				}
+				if len(k.Recurring.DispatchStrategy.AssetForMetric) > 0 && !IsVegaPubkey(k.Recurring.DispatchStrategy.AssetForMetric) {
 					errs.AddForProperty("transfer.kind.dispatch_strategy.asset_for_metric", ErrShouldBeAValidVegaID)
 				}
 				// check that that the metric makes sense for the account type