+++ core/governance/market_cp_restore_test.go
@@ -130,8 +130,9 @@ func createExecutionEngine(t *testing.T, tm time.Time) (*execution.Engine, *gove
 	epochEngine.EXPECT().NotifyOnEpoch(gomock.Any(), gomock.Any()).Times(1)
 
 	bridgeView := amocks.NewMockERC20BridgeView(ctrl)
+	notary := amocks.NewMockNotary(ctrl)
 
-	asset := assets.New(log, assets.NewDefaultConfig(), getNodeWallet(), nil, broker, bridgeView, false)
+	asset := assets.New(log, assets.NewDefaultConfig(), getNodeWallet(), nil, broker, bridgeView, notary, false)
 	marketTracker := execution.NewMarketActivityTracker(log, epochEngine)
 	exec := execution.NewEngine(log, executionConfig, timeService, collateralService, oracleService, broker, statevar, marketTracker, asset)
 	accounts := mocks.NewMockStakingAccounts(ctrl)